feat: move the settings into a separate tab screen
