+++ ExpenseTracker.UI/mobile/app/(tabs)/settings.tsx
@@ -1,0 +1,11 @@
+import ThemedView from "~/components/themed/ThemedView";
+import Settings from "~/components/Settings";
+import { SafeAreaView } from "react-native-safe-area-context";
+
+export default function SettingsScreen() {
+    return (
+        <ThemedView as={SafeAreaView} style={{ flex: 1 }}>
+            <Settings />
+        </ThemedView>
+    );
+}
